Update to service layer
